refactor(settings): group drawer into labeled sections and polish controls
Organize the settings drawer under Appearance / Speech / Suggestions /
About headings, renaming AISettings to SuggestionsSettings to match its
user-facing section. Polish the controls: mute the Rate/Pitch/Volume
slider labels to text.secondary via sx (the color prop no longer resolves
palette paths in MUI v9), and stack the tone radios vertically.
