+++ messages/ar.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "تعليمات مخصصة",
   "aiCustomInstructionsPlaceholder": "مثل: ساخر، مهذّب.",
   "aiCustomInstructionsHelper": "خصّص اقتراحات الذكاء الاصطناعي.",
-  "aiBuiltInSupport": "دعم الذكاء الاصطناعي المدمج",
-  "aiBuiltInSupportHelp": "يتطلب إصدار سطح المكتب من Google Chrome أو Microsoft Edge مع تفعيل العلامات التجريبية.",
-  "aiStatusAvailable": "متاح",
-  "aiStatusUnavailable": "غير متاح",
-  "aiStatusDownloadRequired": "يلزم التنزيل",
-  "aiStatusDownloading": "جارٍ التنزيل…",
   "suggestionsEnable": "تفعيل الاقتراحات",
   "suggestionsDownloading": "جارٍ تنزيل نموذج الذكاء الاصطناعي… {progress}%",
   "suggestionsDownloadingIndeterminate": "جارٍ تنزيل نموذج الذكاء الاصطناعي…",
   "suggestionsUnavailable": "الاقتراحات غير متاحة",
-  "aiFeatureProofreading": "تدقيق لغوي",
-  "aiFeatureRewriting": "إعادة صياغة",
-  "aiFeatureTranslation": "ترجمة",
   "navBack": "رجوع",
   "navHome": "الانتقال إلى الرئيسية",
   "messageBackspace": "مسح",
