+++ messages/bg.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Персонализирани инструкции",
   "aiCustomInstructionsPlaceholder": "напр. Саркастичен, Учтив.",
   "aiCustomInstructionsHelper": "Персонализирайте предложенията на ИИ.",
-  "aiBuiltInSupport": "Вградена поддръжка на ИИ",
-  "aiBuiltInSupportHelp": "Изисква настолната версия на Google Chrome или Microsoft Edge с включени експериментални флагове.",
-  "aiStatusAvailable": "Налично",
-  "aiStatusUnavailable": "Не е налично",
-  "aiStatusDownloadRequired": "Изисква се изтегляне",
-  "aiStatusDownloading": "Изтегляне…",
   "suggestionsEnable": "Включване на предложенията",
   "suggestionsDownloading": "Изтегляне на модела на ИИ… {progress}%",
   "suggestionsDownloadingIndeterminate": "Изтегляне на модела на ИИ…",
   "suggestionsUnavailable": "Предложенията не са налични",
-  "aiFeatureProofreading": "Коректура",
-  "aiFeatureRewriting": "Преписване",
-  "aiFeatureTranslation": "Превод",
   "navBack": "Назад",
   "navHome": "Към началото",
   "messageBackspace": "Изтриване",
