+++ messages/bn.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "কাস্টম নির্দেশনা",
   "aiCustomInstructionsPlaceholder": "যেমন, ব্যঙ্গাত্মক, ভদ্র।",
   "aiCustomInstructionsHelper": "AI পরামর্শ ব্যক্তিগতকৃত করুন।",
-  "aiBuiltInSupport": "অন্তর্নির্মিত AI সমর্থন",
-  "aiBuiltInSupportHelp": "পরীক্ষামূলক ফ্ল্যাগ সক্ষম করা ডেস্কটপ Google Chrome বা Microsoft Edge প্রয়োজন।",
-  "aiStatusAvailable": "উপলব্ধ",
-  "aiStatusUnavailable": "অনুপলব্ধ",
-  "aiStatusDownloadRequired": "ডাউনলোড প্রয়োজন",
-  "aiStatusDownloading": "ডাউনলোড হচ্ছে…",
   "suggestionsEnable": "পরামর্শ চালু করুন",
   "suggestionsDownloading": "AI মডেল ডাউনলোড হচ্ছে… {progress}%",
   "suggestionsDownloadingIndeterminate": "AI মডেল ডাউনলোড হচ্ছে…",
   "suggestionsUnavailable": "পরামর্শ উপলব্ধ নয়",
-  "aiFeatureProofreading": "প্রুফরিডিং",
-  "aiFeatureRewriting": "পুনর্লিখন",
-  "aiFeatureTranslation": "অনুবাদ",
   "navBack": "ফিরে যান",
   "navHome": "হোমে যান",
   "messageBackspace": "মুছুন",
