+++ messages/cs.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Vlastní pokyny",
   "aiCustomInstructionsPlaceholder": "např. Sarkastický, Zdvořilý.",
   "aiCustomInstructionsHelper": "Přizpůsobte si návrhy AI.",
-  "aiBuiltInSupport": "Vestavěná podpora AI",
-  "aiBuiltInSupportHelp": "Vyžaduje desktopový Google Chrome nebo Microsoft Edge se zapnutými experimentálními příznaky.",
-  "aiStatusAvailable": "K dispozici",
-  "aiStatusUnavailable": "Není k dispozici",
-  "aiStatusDownloadRequired": "Vyžaduje stažení",
-  "aiStatusDownloading": "Stahování…",
   "suggestionsEnable": "Zapnout návrhy",
   "suggestionsDownloading": "Stahování modelu AI… {progress} %",
   "suggestionsDownloadingIndeterminate": "Stahování modelu AI…",
   "suggestionsUnavailable": "Návrhy nejsou k dispozici",
-  "aiFeatureProofreading": "Korektura",
-  "aiFeatureRewriting": "Přepis",
-  "aiFeatureTranslation": "Překlad",
   "navBack": "Zpět",
   "navHome": "Přejít domů",
   "messageBackspace": "Smazat",
