+++ messages/da.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Tilpassede instruktioner",
   "aiCustomInstructionsPlaceholder": "f.eks. Sarkastisk, Høflig.",
   "aiCustomInstructionsHelper": "Tilpas AI-forslagene.",
-  "aiBuiltInSupport": "Indbygget AI-understøttelse",
-  "aiBuiltInSupportHelp": "Kræver Google Chrome eller Microsoft Edge på computer med eksperimentelle flag aktiveret.",
-  "aiStatusAvailable": "Tilgængelig",
-  "aiStatusUnavailable": "Ikke tilgængelig",
-  "aiStatusDownloadRequired": "Download påkrævet",
-  "aiStatusDownloading": "Downloader…",
   "suggestionsEnable": "Slå forslag til",
   "suggestionsDownloading": "Downloader AI-model… {progress}%",
   "suggestionsDownloadingIndeterminate": "Downloader AI-model…",
   "suggestionsUnavailable": "Forslag er ikke tilgængelige",
-  "aiFeatureProofreading": "Korrekturlæsning",
-  "aiFeatureRewriting": "Omskrivning",
-  "aiFeatureTranslation": "Oversættelse",
   "navBack": "Tilbage",
   "navHome": "Gå til start",
   "messageBackspace": "Slet",
