+++ messages/de.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Benutzerdefinierte Anweisungen",
   "aiCustomInstructionsPlaceholder": "z. B. Sarkastisch, Höflich.",
   "aiCustomInstructionsHelper": "Personalisiere die KI-Vorschläge.",
-  "aiBuiltInSupport": "Integrierte KI-Unterstützung",
-  "aiBuiltInSupportHelp": "Erfordert die Desktop-Version von Google Chrome oder Microsoft Edge mit aktivierten experimentellen Flags.",
-  "aiStatusAvailable": "Verfügbar",
-  "aiStatusUnavailable": "Nicht verfügbar",
-  "aiStatusDownloadRequired": "Download erforderlich",
-  "aiStatusDownloading": "Wird heruntergeladen…",
   "suggestionsEnable": "Vorschläge aktivieren",
   "suggestionsDownloading": "KI-Modell wird heruntergeladen… {progress}%",
   "suggestionsDownloadingIndeterminate": "KI-Modell wird heruntergeladen…",
   "suggestionsUnavailable": "Vorschläge nicht verfügbar",
-  "aiFeatureProofreading": "Korrektur",
-  "aiFeatureRewriting": "Umformulierung",
-  "aiFeatureTranslation": "Übersetzung",
   "navBack": "Zurück",
   "navHome": "Zur Startseite",
   "messageBackspace": "Löschen",
