+++ messages/el.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Προσαρμοσμένες οδηγίες",
   "aiCustomInstructionsPlaceholder": "π.χ. Σαρκαστικό, Ευγενικό.",
   "aiCustomInstructionsHelper": "Εξατομικεύστε τις προτάσεις του AI.",
-  "aiBuiltInSupport": "Ενσωματωμένη υποστήριξη AI",
-  "aiBuiltInSupportHelp": "Απαιτεί την έκδοση υπολογιστή του Google Chrome ή του Microsoft Edge με ενεργοποιημένες πειραματικές σημαίες.",
-  "aiStatusAvailable": "Διαθέσιμο",
-  "aiStatusUnavailable": "Μη διαθέσιμο",
-  "aiStatusDownloadRequired": "Απαιτείται λήψη",
-  "aiStatusDownloading": "Λήψη…",
   "suggestionsEnable": "Ενεργοποίηση προτάσεων",
   "suggestionsDownloading": "Λήψη μοντέλου AI… {progress}%",
   "suggestionsDownloadingIndeterminate": "Λήψη μοντέλου AI…",
   "suggestionsUnavailable": "Οι προτάσεις δεν είναι διαθέσιμες",
-  "aiFeatureProofreading": "Διόρθωση",
-  "aiFeatureRewriting": "Αναδιατύπωση",
-  "aiFeatureTranslation": "Μετάφραση",
   "navBack": "Πίσω",
   "navHome": "Μετάβαση στην αρχική",
   "messageBackspace": "Διαγραφή",
