+++ messages/en.json
@@ -68,6 +68,10 @@
   "settingsTitle": "Settings",
   "settingsOpen": "Open settings",
   "settingsClose": "Close settings",
+  "settingsSectionAppearance": "Appearance",
+  "settingsSectionSpeech": "Speech",
+  "settingsSectionSuggestions": "Suggestions",
+  "settingsSectionAbout": "About",
   "themeLabel": "Theme",
   "themeSystem": "System",
   "themeLight": "Light",
@@ -85,19 +89,10 @@
   "aiCustomInstructions": "Custom instructions",
   "aiCustomInstructionsPlaceholder": "e.g., sarcastic, polite",
   "aiCustomInstructionsHelper": "Personalize AI suggestions.",
-  "aiBuiltInSupport": "Built-in AI support",
-  "aiBuiltInSupportHelp": "Requires desktop Google Chrome or Microsoft Edge with experimental flags enabled.",
-  "aiStatusAvailable": "Available",
-  "aiStatusUnavailable": "Unavailable",
-  "aiStatusDownloadRequired": "Download required",
-  "aiStatusDownloading": "Downloading...",
   "suggestionsEnable": "Enable suggestions",
   "suggestionsDownloading": "Downloading AI model... {progress}%",
   "suggestionsDownloadingIndeterminate": "Downloading AI model...",
   "suggestionsUnavailable": "Suggestions unavailable",
-  "aiFeatureProofreading": "Proofreading",
-  "aiFeatureRewriting": "Rewriting",
-  "aiFeatureTranslation": "Translation",
   "navBack": "Go back",
   "navHome": "Go home",
   "messageBackspace": "Backspace",
