+++ messages/es.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Instrucciones personalizadas",
   "aiCustomInstructionsPlaceholder": "p. ej., Sarcástico, Educado.",
   "aiCustomInstructionsHelper": "Personaliza las sugerencias de IA.",
-  "aiBuiltInSupport": "Compatibilidad con IA integrada",
-  "aiBuiltInSupportHelp": "Requiere Google Chrome o Microsoft Edge para escritorio con indicadores experimentales habilitados.",
-  "aiStatusAvailable": "Disponible",
-  "aiStatusUnavailable": "No disponible",
-  "aiStatusDownloadRequired": "Se requiere descarga",
-  "aiStatusDownloading": "Descargando…",
   "suggestionsEnable": "Activar sugerencias",
   "suggestionsDownloading": "Descargando modelo de IA… {progress}%",
   "suggestionsDownloadingIndeterminate": "Descargando modelo de IA…",
   "suggestionsUnavailable": "Sugerencias no disponibles",
-  "aiFeatureProofreading": "Corrección",
-  "aiFeatureRewriting": "Reescritura",
-  "aiFeatureTranslation": "Traducción",
   "navBack": "Volver",
   "navHome": "Ir al inicio",
   "messageBackspace": "Borrar",
