+++ messages/fi.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Mukautetut ohjeet",
   "aiCustomInstructionsPlaceholder": "esim. Sarkastinen, Kohtelias.",
   "aiCustomInstructionsHelper": "Mukauta tekoälyn ehdotuksia.",
-  "aiBuiltInSupport": "Sisäänrakennettu tekoälytuki",
-  "aiBuiltInSupportHelp": "Vaatii Google Chromen tai Microsoft Edgen työpöytäversion, jossa kokeelliset liput ovat käytössä.",
-  "aiStatusAvailable": "Saatavilla",
-  "aiStatusUnavailable": "Ei saatavilla",
-  "aiStatusDownloadRequired": "Lataus vaaditaan",
-  "aiStatusDownloading": "Ladataan…",
   "suggestionsEnable": "Ota ehdotukset käyttöön",
   "suggestionsDownloading": "Ladataan tekoälymallia… {progress} %",
   "suggestionsDownloadingIndeterminate": "Ladataan tekoälymallia…",
   "suggestionsUnavailable": "Ehdotukset eivät ole saatavilla",
-  "aiFeatureProofreading": "Oikoluku",
-  "aiFeatureRewriting": "Uudelleenkirjoitus",
-  "aiFeatureTranslation": "Käännös",
   "navBack": "Takaisin",
   "navHome": "Siirry etusivulle",
   "messageBackspace": "Poista",
