+++ messages/fr.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Instructions personnalisées",
   "aiCustomInstructionsPlaceholder": "p. ex. Sarcastique, Poli.",
   "aiCustomInstructionsHelper": "Personnalisez les suggestions de l'IA.",
-  "aiBuiltInSupport": "Prise en charge de l'IA intégrée",
-  "aiBuiltInSupportHelp": "Nécessite la version de bureau de Google Chrome ou Microsoft Edge avec les indicateurs expérimentaux activés.",
-  "aiStatusAvailable": "Disponible",
-  "aiStatusUnavailable": "Indisponible",
-  "aiStatusDownloadRequired": "Téléchargement requis",
-  "aiStatusDownloading": "Téléchargement…",
   "suggestionsEnable": "Activer les suggestions",
   "suggestionsDownloading": "Téléchargement du modèle d'IA… {progress} %",
   "suggestionsDownloadingIndeterminate": "Téléchargement du modèle d'IA…",
   "suggestionsUnavailable": "Suggestions indisponibles",
-  "aiFeatureProofreading": "Correction",
-  "aiFeatureRewriting": "Réécriture",
-  "aiFeatureTranslation": "Traduction",
   "navBack": "Retour",
   "navHome": "Aller à l'accueil",
   "messageBackspace": "Effacer",
