+++ messages/he.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "הוראות מותאמות אישית",
   "aiCustomInstructionsPlaceholder": "למשל: ציני, מנומס.",
   "aiCustomInstructionsHelper": "התאמה אישית של הצעות ה-AI.",
-  "aiBuiltInSupport": "תמיכה ב-AI מובנה",
-  "aiBuiltInSupportHelp": "נדרשת גרסת שולחן העבודה של Google Chrome או Microsoft Edge עם דגלים ניסיוניים מופעלים.",
-  "aiStatusAvailable": "זמין",
-  "aiStatusUnavailable": "לא זמין",
-  "aiStatusDownloadRequired": "נדרשת הורדה",
-  "aiStatusDownloading": "הורדה…",
   "suggestionsEnable": "הפעלת הצעות",
   "suggestionsDownloading": "הורדת מודל ה-AI… {progress}%",
   "suggestionsDownloadingIndeterminate": "הורדת מודל ה-AI…",
   "suggestionsUnavailable": "ההצעות אינן זמינות",
-  "aiFeatureProofreading": "הגהה",
-  "aiFeatureRewriting": "שכתוב",
-  "aiFeatureTranslation": "תרגום",
   "navBack": "חזרה",
   "navHome": "לעמוד הבית",
   "messageBackspace": "מחיקה",
