+++ messages/hi.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "कस्टम निर्देश",
   "aiCustomInstructionsPlaceholder": "जैसे, व्यंग्यात्मक, विनम्र।",
   "aiCustomInstructionsHelper": "AI सुझावों को वैयक्तिकृत करें।",
-  "aiBuiltInSupport": "अंतर्निहित AI समर्थन",
-  "aiBuiltInSupportHelp": "डेस्कटॉप Google Chrome या Microsoft Edge आवश्यक है, जिसमें प्रायोगिक फ़्लैग सक्षम हों।",
-  "aiStatusAvailable": "उपलब्ध",
-  "aiStatusUnavailable": "अनुपलब्ध",
-  "aiStatusDownloadRequired": "डाउनलोड आवश्यक है",
-  "aiStatusDownloading": "डाउनलोड हो रहा है…",
   "suggestionsEnable": "सुझाव चालू करें",
   "suggestionsDownloading": "AI मॉडल डाउनलोड हो रहा है… {progress}%",
   "suggestionsDownloadingIndeterminate": "AI मॉडल डाउनलोड हो रहा है…",
   "suggestionsUnavailable": "सुझाव उपलब्ध नहीं हैं",
-  "aiFeatureProofreading": "प्रूफ़रीडिंग",
-  "aiFeatureRewriting": "पुनर्लेखन",
-  "aiFeatureTranslation": "अनुवाद",
   "navBack": "वापस जाएँ",
   "navHome": "होम पर जाएँ",
   "messageBackspace": "मिटाएं",
