+++ messages/hr.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Prilagođene upute",
   "aiCustomInstructionsPlaceholder": "npr. Sarkastičan, Pristojan.",
   "aiCustomInstructionsHelper": "Prilagodite prijedloge umjetne inteligencije.",
-  "aiBuiltInSupport": "Ugrađena podrška za umjetnu inteligenciju",
-  "aiBuiltInSupportHelp": "Zahtijeva stolnu verziju preglednika Google Chrome ili Microsoft Edge s omogućenim eksperimentalnim zastavicama.",
-  "aiStatusAvailable": "Dostupno",
-  "aiStatusUnavailable": "Nedostupno",
-  "aiStatusDownloadRequired": "Potrebno je preuzimanje",
-  "aiStatusDownloading": "Preuzimanje…",
   "suggestionsEnable": "Uključi prijedloge",
   "suggestionsDownloading": "Preuzimanje AI modela… {progress} %",
   "suggestionsDownloadingIndeterminate": "Preuzimanje AI modela…",
   "suggestionsUnavailable": "Prijedlozi nisu dostupni",
-  "aiFeatureProofreading": "Lektura",
-  "aiFeatureRewriting": "Preoblikovanje",
-  "aiFeatureTranslation": "Prijevod",
   "navBack": "Natrag",
   "navHome": "Idi na početnu",
   "messageBackspace": "Obriši",
