+++ messages/hu.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Egyéni utasítások",
   "aiCustomInstructionsPlaceholder": "pl. Szarkasztikus, Udvarias.",
   "aiCustomInstructionsHelper": "Személyre szabhatod az MI-javaslatokat.",
-  "aiBuiltInSupport": "Beépített MI-támogatás",
-  "aiBuiltInSupportHelp": "Az asztali Google Chrome vagy Microsoft Edge szükséges, engedélyezett kísérleti jelzőkkel.",
-  "aiStatusAvailable": "Elérhető",
-  "aiStatusUnavailable": "Nem érhető el",
-  "aiStatusDownloadRequired": "Letöltés szükséges",
-  "aiStatusDownloading": "Letöltés…",
   "suggestionsEnable": "Javaslatok bekapcsolása",
   "suggestionsDownloading": "MI-modell letöltése… {progress}%",
   "suggestionsDownloadingIndeterminate": "MI-modell letöltése…",
   "suggestionsUnavailable": "A javaslatok nem érhetők el",
-  "aiFeatureProofreading": "Korrektúra",
-  "aiFeatureRewriting": "Átfogalmazás",
-  "aiFeatureTranslation": "Fordítás",
   "navBack": "Vissza",
   "navHome": "Ugrás a kezdőlapra",
   "messageBackspace": "Törlés",
