+++ messages/id.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Instruksi khusus",
   "aiCustomInstructionsPlaceholder": "mis. Sarkastis, Sopan.",
   "aiCustomInstructionsHelper": "Personalisasikan saran AI.",
-  "aiBuiltInSupport": "Dukungan AI bawaan",
-  "aiBuiltInSupportHelp": "Memerlukan Google Chrome atau Microsoft Edge versi desktop dengan flag eksperimental diaktifkan.",
-  "aiStatusAvailable": "Tersedia",
-  "aiStatusUnavailable": "Tidak tersedia",
-  "aiStatusDownloadRequired": "Perlu diunduh",
-  "aiStatusDownloading": "Mengunduh…",
   "suggestionsEnable": "Aktifkan saran",
   "suggestionsDownloading": "Mengunduh model AI… {progress}%",
   "suggestionsDownloadingIndeterminate": "Mengunduh model AI…",
   "suggestionsUnavailable": "Saran tidak tersedia",
-  "aiFeatureProofreading": "Koreksi",
-  "aiFeatureRewriting": "Penulisan ulang",
-  "aiFeatureTranslation": "Terjemahan",
   "navBack": "Kembali",
   "navHome": "Ke beranda",
   "messageBackspace": "Hapus",
