+++ messages/it.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Istruzioni personalizzate",
   "aiCustomInstructionsPlaceholder": "ad es. Sarcastico, Cortese.",
   "aiCustomInstructionsHelper": "Personalizza i suggerimenti dell'IA.",
-  "aiBuiltInSupport": "Supporto IA integrata",
-  "aiBuiltInSupportHelp": "Richiede Google Chrome o Microsoft Edge per desktop con i flag sperimentali abilitati.",
-  "aiStatusAvailable": "Disponibile",
-  "aiStatusUnavailable": "Non disponibile",
-  "aiStatusDownloadRequired": "Download richiesto",
-  "aiStatusDownloading": "Download…",
   "suggestionsEnable": "Attiva i suggerimenti",
   "suggestionsDownloading": "Download del modello IA… {progress}%",
   "suggestionsDownloadingIndeterminate": "Download del modello IA…",
   "suggestionsUnavailable": "Suggerimenti non disponibili",
-  "aiFeatureProofreading": "Correzione",
-  "aiFeatureRewriting": "Riscrittura",
-  "aiFeatureTranslation": "Traduzione",
   "navBack": "Indietro",
   "navHome": "Vai alla home",
   "messageBackspace": "Cancella",
