+++ messages/ja.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "カスタム指示",
   "aiCustomInstructionsPlaceholder": "例：皮肉っぽく、丁寧に。",
   "aiCustomInstructionsHelper": "AI の提案をカスタマイズします。",
-  "aiBuiltInSupport": "組み込み AI のサポート",
-  "aiBuiltInSupportHelp": "デスクトップ版の Google Chrome または Microsoft Edge で試験運用版のフラグを有効にする必要があります。",
-  "aiStatusAvailable": "利用可能",
-  "aiStatusUnavailable": "利用不可",
-  "aiStatusDownloadRequired": "ダウンロードが必要",
-  "aiStatusDownloading": "ダウンロード中…",
   "suggestionsEnable": "提案を有効にする",
   "suggestionsDownloading": "AI モデルをダウンロード中… {progress}%",
   "suggestionsDownloadingIndeterminate": "AI モデルをダウンロード中…",
   "suggestionsUnavailable": "提案は利用できません",
-  "aiFeatureProofreading": "校正",
-  "aiFeatureRewriting": "リライト",
-  "aiFeatureTranslation": "翻訳",
   "navBack": "戻る",
   "navHome": "ホームに戻る",
   "messageBackspace": "消去",
