+++ messages/kn.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "ಕಸ್ಟಮ್ ಸೂಚನೆಗಳು",
   "aiCustomInstructionsPlaceholder": "ಉದಾ., ವ್ಯಂಗ್ಯ, ಸಭ್ಯ.",
   "aiCustomInstructionsHelper": "AI ಸಲಹೆಗಳನ್ನು ವೈಯಕ್ತೀಕರಿಸಿ.",
-  "aiBuiltInSupport": "ಅಂತರ್ನಿರ್ಮಿತ AI ಬೆಂಬಲ",
-  "aiBuiltInSupportHelp": "ಡೆಸ್ಕ್‌ಟಾಪ್ Google Chrome ಅಥವಾ Microsoft Edge ಅಗತ್ಯವಿದೆ, ಪ್ರಾಯೋಗಿಕ ಫ್ಲ್ಯಾಗ್‌ಗಳನ್ನು ಸಕ್ರಿಯಗೊಳಿಸಿರಬೇಕು.",
-  "aiStatusAvailable": "ಲಭ್ಯವಿದೆ",
-  "aiStatusUnavailable": "ಲಭ್ಯವಿಲ್ಲ",
-  "aiStatusDownloadRequired": "ಡೌನ್‌ಲೋಡ್ ಅಗತ್ಯವಿದೆ",
-  "aiStatusDownloading": "ಡೌನ್‌ಲೋಡ್ ಆಗುತ್ತಿದೆ…",
   "suggestionsEnable": "ಸಲಹೆಗಳನ್ನು ಸಕ್ರಿಯಗೊಳಿಸಿ",
   "suggestionsDownloading": "AI ಮಾದರಿ ಡೌನ್‌ಲೋಡ್ ಆಗುತ್ತಿದೆ… {progress}%",
   "suggestionsDownloadingIndeterminate": "AI ಮಾದರಿ ಡೌನ್‌ಲೋಡ್ ಆಗುತ್ತಿದೆ…",
   "suggestionsUnavailable": "ಸಲಹೆಗಳು ಲಭ್ಯವಿಲ್ಲ",
-  "aiFeatureProofreading": "ಪ್ರೂಫ್‌ರೀಡಿಂಗ್",
-  "aiFeatureRewriting": "ಮರುಬರವಣಿಗೆ",
-  "aiFeatureTranslation": "ಅನುವಾದ",
   "navBack": "ಹಿಂದಕ್ಕೆ",
   "navHome": "ಮುಖಪುಟಕ್ಕೆ ಹೋಗಿ",
   "messageBackspace": "ಅಳಿಸು",
