+++ messages/ko.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "맞춤 지침",
   "aiCustomInstructionsPlaceholder": "예: 냉소적으로, 공손하게.",
   "aiCustomInstructionsHelper": "AI 제안을 개인화하세요.",
-  "aiBuiltInSupport": "내장 AI 지원",
-  "aiBuiltInSupportHelp": "데스크톱용 Google Chrome 또는 Microsoft Edge에서 실험용 플래그를 사용 설정해야 합니다.",
-  "aiStatusAvailable": "사용 가능",
-  "aiStatusUnavailable": "사용 불가",
-  "aiStatusDownloadRequired": "다운로드 필요",
-  "aiStatusDownloading": "다운로드 중…",
   "suggestionsEnable": "제안 켜기",
   "suggestionsDownloading": "AI 모델 다운로드 중… {progress}%",
   "suggestionsDownloadingIndeterminate": "AI 모델 다운로드 중…",
   "suggestionsUnavailable": "제안을 사용할 수 없음",
-  "aiFeatureProofreading": "교정",
-  "aiFeatureRewriting": "다시 쓰기",
-  "aiFeatureTranslation": "번역",
   "navBack": "뒤로",
   "navHome": "홈으로",
   "messageBackspace": "지우기",
