+++ messages/nl.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Aangepaste instructies",
   "aiCustomInstructionsPlaceholder": "bijv. Sarcastisch, Beleefd.",
   "aiCustomInstructionsHelper": "Personaliseer de AI-suggesties.",
-  "aiBuiltInSupport": "Ingebouwde AI-ondersteuning",
-  "aiBuiltInSupportHelp": "Vereist de desktopversie van Google Chrome of Microsoft Edge met experimentele vlaggen ingeschakeld.",
-  "aiStatusAvailable": "Beschikbaar",
-  "aiStatusUnavailable": "Niet beschikbaar",
-  "aiStatusDownloadRequired": "Download vereist",
-  "aiStatusDownloading": "Downloaden…",
   "suggestionsEnable": "Suggesties inschakelen",
   "suggestionsDownloading": "AI-model downloaden… {progress}%",
   "suggestionsDownloadingIndeterminate": "AI-model downloaden…",
   "suggestionsUnavailable": "Suggesties niet beschikbaar",
-  "aiFeatureProofreading": "Correctie",
-  "aiFeatureRewriting": "Herschrijven",
-  "aiFeatureTranslation": "Vertaling",
   "navBack": "Terug",
   "navHome": "Naar start",
   "messageBackspace": "Wissen",
