+++ messages/pl.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Instrukcje niestandardowe",
   "aiCustomInstructionsPlaceholder": "np. Sarkastyczny, Uprzejmy.",
   "aiCustomInstructionsHelper": "Spersonalizuj sugestie AI.",
-  "aiBuiltInSupport": "Wbudowana obsługa AI",
-  "aiBuiltInSupportHelp": "Wymaga przeglądarki Google Chrome lub Microsoft Edge na komputery z włączonymi flagami eksperymentalnymi.",
-  "aiStatusAvailable": "Dostępne",
-  "aiStatusUnavailable": "Niedostępne",
-  "aiStatusDownloadRequired": "Wymagane pobranie",
-  "aiStatusDownloading": "Pobieranie…",
   "suggestionsEnable": "Włącz sugestie",
   "suggestionsDownloading": "Pobieranie modelu AI… {progress}%",
   "suggestionsDownloadingIndeterminate": "Pobieranie modelu AI…",
   "suggestionsUnavailable": "Sugestie są niedostępne",
-  "aiFeatureProofreading": "Korekta",
-  "aiFeatureRewriting": "Przeredagowanie",
-  "aiFeatureTranslation": "Tłumaczenie",
   "navBack": "Wstecz",
   "navHome": "Przejdź do strony głównej",
   "messageBackspace": "Usuń",
