+++ messages/pt.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Instruções personalizadas",
   "aiCustomInstructionsPlaceholder": "ex.: Sarcástico, Educado.",
   "aiCustomInstructionsHelper": "Personalize as sugestões da IA.",
-  "aiBuiltInSupport": "Suporte a IA integrada",
-  "aiBuiltInSupportHelp": "Requer o Google Chrome ou Microsoft Edge para computador com sinalizadores experimentais ativados.",
-  "aiStatusAvailable": "Disponível",
-  "aiStatusUnavailable": "Indisponível",
-  "aiStatusDownloadRequired": "Download necessário",
-  "aiStatusDownloading": "Baixando…",
   "suggestionsEnable": "Ativar sugestões",
   "suggestionsDownloading": "Baixando modelo de IA… {progress}%",
   "suggestionsDownloadingIndeterminate": "Baixando modelo de IA…",
   "suggestionsUnavailable": "Sugestões indisponíveis",
-  "aiFeatureProofreading": "Revisão",
-  "aiFeatureRewriting": "Reescrita",
-  "aiFeatureTranslation": "Tradução",
   "navBack": "Voltar",
   "navHome": "Ir para o início",
   "messageBackspace": "Apagar",
