+++ messages/ro.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Instrucțiuni personalizate",
   "aiCustomInstructionsPlaceholder": "ex. Sarcastic, Politicos.",
   "aiCustomInstructionsHelper": "Personalizează sugestiile AI.",
-  "aiBuiltInSupport": "Suport AI integrat",
-  "aiBuiltInSupportHelp": "Necesită versiunea desktop a Google Chrome sau Microsoft Edge cu indicatoarele experimentale activate.",
-  "aiStatusAvailable": "Disponibil",
-  "aiStatusUnavailable": "Indisponibil",
-  "aiStatusDownloadRequired": "Necesită descărcare",
-  "aiStatusDownloading": "Se descarcă…",
   "suggestionsEnable": "Activează sugestiile",
   "suggestionsDownloading": "Se descarcă modelul AI… {progress}%",
   "suggestionsDownloadingIndeterminate": "Se descarcă modelul AI…",
   "suggestionsUnavailable": "Sugestiile nu sunt disponibile",
-  "aiFeatureProofreading": "Corectură",
-  "aiFeatureRewriting": "Rescriere",
-  "aiFeatureTranslation": "Traducere",
   "navBack": "Înapoi",
   "navHome": "Mergi la pagina principală",
   "messageBackspace": "Șterge",
