+++ messages/ru.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Пользовательские инструкции",
   "aiCustomInstructionsPlaceholder": "напр. Саркастичный, Вежливый.",
   "aiCustomInstructionsHelper": "Персонализируйте предложения ИИ.",
-  "aiBuiltInSupport": "Встроенная поддержка ИИ",
-  "aiBuiltInSupportHelp": "Требуется настольная версия Google Chrome или Microsoft Edge с включёнными экспериментальными флагами.",
-  "aiStatusAvailable": "Доступно",
-  "aiStatusUnavailable": "Недоступно",
-  "aiStatusDownloadRequired": "Требуется загрузка",
-  "aiStatusDownloading": "Загрузка…",
   "suggestionsEnable": "Включить предложения",
   "suggestionsDownloading": "Загрузка модели ИИ… {progress}%",
   "suggestionsDownloadingIndeterminate": "Загрузка модели ИИ…",
   "suggestionsUnavailable": "Предложения недоступны",
-  "aiFeatureProofreading": "Корректура",
-  "aiFeatureRewriting": "Перефразирование",
-  "aiFeatureTranslation": "Перевод",
   "navBack": "Назад",
   "navHome": "На главную",
   "messageBackspace": "Стереть",
