+++ messages/sk.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Vlastné pokyny",
   "aiCustomInstructionsPlaceholder": "napr. Sarkastický, Zdvorilý.",
   "aiCustomInstructionsHelper": "Prispôsobte si návrhy AI.",
-  "aiBuiltInSupport": "Vstavaná podpora AI",
-  "aiBuiltInSupportHelp": "Vyžaduje počítačovú verziu Google Chrome alebo Microsoft Edge so zapnutými experimentálnymi príznakmi.",
-  "aiStatusAvailable": "K dispozícii",
-  "aiStatusUnavailable": "Nie je k dispozícii",
-  "aiStatusDownloadRequired": "Vyžaduje sa stiahnutie",
-  "aiStatusDownloading": "Sťahovanie…",
   "suggestionsEnable": "Zapnúť návrhy",
   "suggestionsDownloading": "Sťahovanie modelu AI… {progress} %",
   "suggestionsDownloadingIndeterminate": "Sťahovanie modelu AI…",
   "suggestionsUnavailable": "Návrhy nie sú k dispozícii",
-  "aiFeatureProofreading": "Korektúra",
-  "aiFeatureRewriting": "Prepis",
-  "aiFeatureTranslation": "Preklad",
   "navBack": "Späť",
   "navHome": "Prejsť domov",
   "messageBackspace": "Vymazať",
