+++ messages/sl.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Navodila po meri",
   "aiCustomInstructionsPlaceholder": "npr. Sarkastičen, Vljuden.",
   "aiCustomInstructionsHelper": "Prilagodite predloge UI.",
-  "aiBuiltInSupport": "Vgrajena podpora za UI",
-  "aiBuiltInSupportHelp": "Zahteva namizno različico Google Chrome ali Microsoft Edge z omogočenimi eksperimentalnimi zastavicami.",
-  "aiStatusAvailable": "Na voljo",
-  "aiStatusUnavailable": "Ni na voljo",
-  "aiStatusDownloadRequired": "Potreben je prenos",
-  "aiStatusDownloading": "Prenašanje…",
   "suggestionsEnable": "Vklopi predloge",
   "suggestionsDownloading": "Prenašanje modela UI… {progress} %",
   "suggestionsDownloadingIndeterminate": "Prenašanje modela UI…",
   "suggestionsUnavailable": "Predlogi niso na voljo",
-  "aiFeatureProofreading": "Lektoriranje",
-  "aiFeatureRewriting": "Preoblikovanje",
-  "aiFeatureTranslation": "Prevajanje",
   "navBack": "Nazaj",
   "navHome": "Pojdi domov",
   "messageBackspace": "Briši",
