+++ messages/sv.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Anpassade instruktioner",
   "aiCustomInstructionsPlaceholder": "t.ex. Sarkastisk, Artig.",
   "aiCustomInstructionsHelper": "Anpassa AI-förslagen.",
-  "aiBuiltInSupport": "Inbyggt AI-stöd",
-  "aiBuiltInSupportHelp": "Kräver Google Chrome eller Microsoft Edge på datorn med experimentella flaggor aktiverade.",
-  "aiStatusAvailable": "Tillgänglig",
-  "aiStatusUnavailable": "Inte tillgänglig",
-  "aiStatusDownloadRequired": "Nedladdning krävs",
-  "aiStatusDownloading": "Laddar ner…",
   "suggestionsEnable": "Aktivera förslag",
   "suggestionsDownloading": "Laddar ner AI-modell… {progress}%",
   "suggestionsDownloadingIndeterminate": "Laddar ner AI-modell…",
   "suggestionsUnavailable": "Förslag är inte tillgängliga",
-  "aiFeatureProofreading": "Korrekturläsning",
-  "aiFeatureRewriting": "Omskrivning",
-  "aiFeatureTranslation": "Översättning",
   "navBack": "Tillbaka",
   "navHome": "Till startsidan",
   "messageBackspace": "Radera",
