+++ messages/ta.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "தனிப்பயன் வழிமுறைகள்",
   "aiCustomInstructionsPlaceholder": "எ.கா., கிண்டலான, பணிவான.",
   "aiCustomInstructionsHelper": "AI பரிந்துரைகளைத் தனிப்பயனாக்கு.",
-  "aiBuiltInSupport": "உள்ளமைந்த AI ஆதரவு",
-  "aiBuiltInSupportHelp": "சோதனை ஃபிளாக்குகள் இயக்கப்பட்ட டெஸ்க்டாப் Google Chrome அல்லது Microsoft Edge தேவை.",
-  "aiStatusAvailable": "கிடைக்கிறது",
-  "aiStatusUnavailable": "கிடைக்கவில்லை",
-  "aiStatusDownloadRequired": "பதிவிறக்கம் தேவை",
-  "aiStatusDownloading": "பதிவிறக்கப்படுகிறது…",
   "suggestionsEnable": "பரிந்துரைகளை இயக்கு",
   "suggestionsDownloading": "AI மாடல் பதிவிறக்கப்படுகிறது… {progress}%",
   "suggestionsDownloadingIndeterminate": "AI மாடல் பதிவிறக்கப்படுகிறது…",
   "suggestionsUnavailable": "பரிந்துரைகள் கிடைக்கவில்லை",
-  "aiFeatureProofreading": "மெய்ப்புத் திருத்தம்",
-  "aiFeatureRewriting": "மறுஎழுத்து",
-  "aiFeatureTranslation": "மொழிபெயர்ப்பு",
   "navBack": "பின்செல்",
   "navHome": "முகப்புக்குச் செல்",
   "messageBackspace": "அழி",
