+++ messages/te.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "అనుకూల సూచనలు",
   "aiCustomInstructionsPlaceholder": "ఉదా., వ్యంగ్యంగా, మర్యాదగా.",
   "aiCustomInstructionsHelper": "AI సూచనలను వ్యక్తిగతీకరించండి.",
-  "aiBuiltInSupport": "అంతర్నిర్మిత AI మద్దతు",
-  "aiBuiltInSupportHelp": "ప్రయోగాత్మక ఫ్లాగ్‌లు ప్రారంభించిన డెస్క్‌టాప్ Google Chrome లేదా Microsoft Edge అవసరం.",
-  "aiStatusAvailable": "అందుబాటులో ఉంది",
-  "aiStatusUnavailable": "అందుబాటులో లేదు",
-  "aiStatusDownloadRequired": "డౌన్‌లోడ్ అవసరం",
-  "aiStatusDownloading": "డౌన్‌లోడ్ అవుతోంది…",
   "suggestionsEnable": "సూచనలను ప్రారంభించండి",
   "suggestionsDownloading": "AI మోడల్ డౌన్‌లోడ్ అవుతోంది… {progress}%",
   "suggestionsDownloadingIndeterminate": "AI మోడల్ డౌన్‌లోడ్ అవుతోంది…",
   "suggestionsUnavailable": "సూచనలు అందుబాటులో లేవు",
-  "aiFeatureProofreading": "ప్రూఫ్‌రీడింగ్",
-  "aiFeatureRewriting": "తిరిగి రాయడం",
-  "aiFeatureTranslation": "అనువాదం",
   "navBack": "వెనుకకు",
   "navHome": "హోమ్‌కు వెళ్లు",
   "messageBackspace": "తుడిచివేయి",
