+++ messages/th.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "คำสั่งที่กำหนดเอง",
   "aiCustomInstructionsPlaceholder": "เช่น ประชดประชัน สุภาพ",
   "aiCustomInstructionsHelper": "ปรับแต่งคำแนะนำของ AI ให้เป็นส่วนตัว",
-  "aiBuiltInSupport": "การรองรับ AI ในตัว",
-  "aiBuiltInSupportHelp": "ต้องใช้ Google Chrome หรือ Microsoft Edge บนเดสก์ท็อปโดยเปิดใช้งานแฟล็กทดลอง",
-  "aiStatusAvailable": "พร้อมใช้งาน",
-  "aiStatusUnavailable": "ไม่พร้อมใช้งาน",
-  "aiStatusDownloadRequired": "ต้องดาวน์โหลดก่อน",
-  "aiStatusDownloading": "กำลังดาวน์โหลด…",
   "suggestionsEnable": "เปิดใช้คำแนะนำ",
   "suggestionsDownloading": "กำลังดาวน์โหลดโมเดล AI… {progress}%",
   "suggestionsDownloadingIndeterminate": "กำลังดาวน์โหลดโมเดล AI…",
   "suggestionsUnavailable": "คำแนะนำไม่พร้อมใช้งาน",
-  "aiFeatureProofreading": "การตรวจทาน",
-  "aiFeatureRewriting": "การเขียนใหม่",
-  "aiFeatureTranslation": "การแปล",
   "navBack": "กลับ",
   "navHome": "ไปที่หน้าแรก",
   "messageBackspace": "ลบ",
