+++ messages/tr.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Özel talimatlar",
   "aiCustomInstructionsPlaceholder": "ör. Alaycı, Kibar.",
   "aiCustomInstructionsHelper": "Yapay zeka önerilerini kişiselleştirin.",
-  "aiBuiltInSupport": "Yerleşik yapay zeka desteği",
-  "aiBuiltInSupportHelp": "Masaüstü Google Chrome veya Microsoft Edge'de deneysel işaretlerin etkinleştirilmesi gerekir.",
-  "aiStatusAvailable": "Kullanılabilir",
-  "aiStatusUnavailable": "Kullanılamıyor",
-  "aiStatusDownloadRequired": "İndirme gerekli",
-  "aiStatusDownloading": "İndiriliyor…",
   "suggestionsEnable": "Önerileri aç",
   "suggestionsDownloading": "Yapay zeka modeli indiriliyor… %{progress}",
   "suggestionsDownloadingIndeterminate": "Yapay zeka modeli indiriliyor…",
   "suggestionsUnavailable": "Öneriler kullanılamıyor",
-  "aiFeatureProofreading": "Düzeltme",
-  "aiFeatureRewriting": "Yeniden yazma",
-  "aiFeatureTranslation": "Çeviri",
   "navBack": "Geri",
   "navHome": "Ana sayfaya git",
   "messageBackspace": "Geri sil",
