+++ messages/uk.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Власні інструкції",
   "aiCustomInstructionsPlaceholder": "напр. Саркастичний, Ввічливий.",
   "aiCustomInstructionsHelper": "Персоналізуйте пропозиції ШІ.",
-  "aiBuiltInSupport": "Вбудована підтримка ШІ",
-  "aiBuiltInSupportHelp": "Потрібна настільна версія Google Chrome або Microsoft Edge з увімкненими експериментальними прапорцями.",
-  "aiStatusAvailable": "Доступно",
-  "aiStatusUnavailable": "Недоступно",
-  "aiStatusDownloadRequired": "Потрібне завантаження",
-  "aiStatusDownloading": "Завантаження…",
   "suggestionsEnable": "Увімкнути пропозиції",
   "suggestionsDownloading": "Завантаження моделі ШІ… {progress}%",
   "suggestionsDownloadingIndeterminate": "Завантаження моделі ШІ…",
   "suggestionsUnavailable": "Пропозиції недоступні",
-  "aiFeatureProofreading": "Коректура",
-  "aiFeatureRewriting": "Переписування",
-  "aiFeatureTranslation": "Переклад",
   "navBack": "Назад",
   "navHome": "На головну",
   "messageBackspace": "Стерти",
