+++ messages/vi.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "Hướng dẫn tùy chỉnh",
   "aiCustomInstructionsPlaceholder": "ví dụ: Châm biếm, Lịch sự.",
   "aiCustomInstructionsHelper": "Cá nhân hóa gợi ý của AI.",
-  "aiBuiltInSupport": "Hỗ trợ AI tích hợp",
-  "aiBuiltInSupportHelp": "Yêu cầu Google Chrome hoặc Microsoft Edge phiên bản máy tính với cờ thử nghiệm được bật.",
-  "aiStatusAvailable": "Khả dụng",
-  "aiStatusUnavailable": "Không khả dụng",
-  "aiStatusDownloadRequired": "Cần tải xuống",
-  "aiStatusDownloading": "Đang tải xuống…",
   "suggestionsEnable": "Bật gợi ý",
   "suggestionsDownloading": "Đang tải mô hình AI xuống… {progress}%",
   "suggestionsDownloadingIndeterminate": "Đang tải mô hình AI xuống…",
   "suggestionsUnavailable": "Gợi ý không khả dụng",
-  "aiFeatureProofreading": "Hiệu đính",
-  "aiFeatureRewriting": "Viết lại",
-  "aiFeatureTranslation": "Dịch",
   "navBack": "Quay lại",
   "navHome": "Về trang chủ",
   "messageBackspace": "Xóa",
