+++ messages/zh.json
@@ -85,19 +85,10 @@
   "aiCustomInstructions": "自定义指令",
   "aiCustomInstructionsPlaceholder": "例如：讽刺、礼貌。",
   "aiCustomInstructionsHelper": "个性化 AI 建议。",
-  "aiBuiltInSupport": "内置 AI 支持",
-  "aiBuiltInSupportHelp": "需要桌面版 Google Chrome 或 Microsoft Edge 并启用实验性标志。",
-  "aiStatusAvailable": "可用",
-  "aiStatusUnavailable": "不可用",
-  "aiStatusDownloadRequired": "需要下载",
-  "aiStatusDownloading": "正在下载…",
   "suggestionsEnable": "启用建议",
   "suggestionsDownloading": "正在下载 AI 模型… {progress}%",
   "suggestionsDownloadingIndeterminate": "正在下载 AI 模型…",
   "suggestionsUnavailable": "建议不可用",
-  "aiFeatureProofreading": "校对",
-  "aiFeatureRewriting": "改写",
-  "aiFeatureTranslation": "翻译",
   "navBack": "返回",
   "navHome": "返回主页",
   "messageBackspace": "删除",
