+++ src/app/settings/settings-drawer.test.tsx
@@ -1,11 +1,20 @@
 import { AppProviders } from "@shared/providers/app-providers";
 import { expectNoA11yViolations } from "@shared/testing/axe";
+import {
+  stubProofreader,
+  stubRewriter,
+  stubTranslator,
+} from "@shared/testing/built-in-ai";
 import { describe, expect, test, vi } from "vitest";
 import { render } from "vitest-browser-react";
 import { SettingsDrawer } from "./settings-drawer";
 
 describe("SettingsDrawer", () => {
   test("renders every settings section when open, with no a11y violations", async () => {
+    stubProofreader();
+    stubRewriter();
+    stubTranslator();
+
     const screen = await render(
       <AppProviders>
         <SettingsDrawer open onClose={vi.fn()} />
@@ -13,9 +22,17 @@ describe("SettingsDrawer", () => {
     );
 
     await expect.element(screen.getByText("Settings")).toBeInTheDocument();
+
+    await expect
+      .element(screen.getByRole("heading", { name: "Appearance" }))
+      .toBeInTheDocument();
     await expect.element(screen.getByText("Theme")).toBeInTheDocument();
     await expect
       .element(screen.getByRole("combobox", { name: "Language" }))
+      .toBeInTheDocument();
+
+    await expect
+      .element(screen.getByRole("heading", { name: "Speech" }))
       .toBeInTheDocument();
     await expect
       .element(screen.getByRole("combobox", { name: "Voice" }))
@@ -23,8 +40,19 @@ describe("SettingsDrawer", () => {
     await expect
       .element(screen.getByRole("switch", { name: "Highlight while playing" }))
       .toBeInTheDocument();
+
     await expect
-      .element(screen.getByText("Built-in AI support"))
+      .element(screen.getByRole("heading", { name: "Suggestions" }))
+      .toBeInTheDocument();
+    await expect
+      .element(screen.getByRole("textbox", { name: "Custom instructions" }))
+      .toBeInTheDocument();
+
+    await expect
+      .element(screen.getByRole("heading", { name: "About" }))
+      .toBeInTheDocument();
+    await expect
+      .element(screen.getByRole("link", { name: /Source code/i }))
       .toBeInTheDocument();
 
     await expectNoA11yViolations(document.body);
